+++ keyboards/frooastboard/nano/nano.c
@@ -1,4 +1,4 @@
-/* Copyright 2021 frooastside
+/* Copyright 2021-2022 frooastside
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -14,14 +14,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include QMK_KEYBOARD_H
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-	[0] = LAYOUT(
-		KC_Z, KC_X,
-		LT(1, KC_ESC), KC_TILD),
-
-	[1] = LAYOUT(
-		KC_F23, KC_F24,
-		KC_TRNS, RGB_TOG)
-};
+#include "nano.h"
