+++ keyboards/frooastboard/nano/readme.md
@@ -1,19 +1,19 @@
-# Frooastboard
+# Frooastboard Nano
 
-![Frooastboard](https://i.imgur.com/f26O8JLh.jpg)
+![Frooastboard Nano](https://i.imgur.com/f26O8JLh.jpg)
 
 * Keyboard Maintainer: [Frooastside](https://github.com/frooastside)
-* Hardware Supported: Frooastboard
+* Hardware Supported: Frooastboard Nano
 * Hardware Availability: https://github.com/Frooastside/Frooastboard
 
 **Reset Key:** Hold down the key located at `K00`, commonly programmed as ESCAPE while plugging in the keyboard. (If via is used it is the key located at `K10`.)
 
 Make example for this keyboard (after setting up your build environment):
 
-    make frooastboard:default
+    make frooastboard/nano:default
 
 Flashing example for this keyboard:
 
-    make frooastboard:default:flash
+    make frooastboard/nano:default:flash
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
